Keep the CHRS kludge parser inside its buffers
Three faults in code run for every header while an area is listed.
The QP suffix was cut out of the charset name at an offset measured on
the source word, after the value had already been shortened to the
caller's buffer: a first word of 64 or 65 bytes ending in QP read and
wrote past a 64-byte buffer on the stack. The length is now measured
on the copy. A kludge that was exactly "RFC" stepped past its own
terminator looking for the separator after the prefix; the terminator
is checked first. And a charset name carrying a byte above 0x7F went
through toupper() as a negative value - harmless under GCC, which
compiles with unsigned char, but the Win32 table lookups indexed
before the table. The byte is taken as unsigned, in the caller and in
the Win32 inline functions alike.

+++ goldlib/gall/gctype.h
@@ -54,13 +54,16 @@ extern char tl[256], tu[256];
 //  Plain `inline': this block is inside #ifdef __cplusplus, so the
 //  keyword is the right one everywhere, and Borland's compiler will not
 //  take __inline inside an extern "C" block at all.
+//  The cast matches the C macros below: a plain char promoted to int is
+//  negative for anything above 0x7F, and a negative subscript reads
+//  before the table.
 inline int g_tolower(int c)
 {
-    return tl[c];
+    return tl[(uint8_t)c];
 }
 inline int g_toupper(int c)
 {
-    return tu[c];
+    return tu[(uint8_t)c];
 }
 }
 #else
